CourseSettingsButton in it's own component; improve darkmode legibility
